Move ArticleAdmin into bases.py
Use ARTICLE_MODELADMIN_CLASS setting
